\name{setplot}
\alias{setplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Set graphical parameters for  geographical maps so as to have equal ratio for horizontal
and vertical axis. Internal function.}
\description{Set graphical parameters so as to have equal ratio for horizontal
and vertical axis.}
\usage{
setplot(xdata, ydata, pretty.call = TRUE, maxdim, axes = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xdata}{A vector of x coordinates}
  \item{ydata}{A vector of y coordinates}
  \item{pretty.call}{ TODO }
  \item{maxdim}{TODO }
  \item{axes}{ TODO}
}

\value{Reset graphical parameters}

\author{Unknown}

\keyword{}

\examples{
#library(Geneclust)

}

